# specs/001-boq-price-forecast/contracts/api-openapi.yaml
openapi: 3.1.0
info:
  title: BoQ Price Forecast API
  version: 1.0.0
  description: |
    REST API for AI-powered BoQ price adjustment and forecasting.
    Supports Excel upload, weight attribution, price recalculation, forecasting, and defense document generation.
  contact:
    name: BoQ Price Forecast Team
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.boqprice.example.com
    description: Production server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
        params:
          type: object
      required: [detail]
...
    PriceCalculationRequest:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        item_ids:
          type: array
          items:
            type: string
            format: uuid
        custom_multipliers:
          type: object
          properties:
            risk_buffer:
              type: number
              format: double
            payment_terms:
              type: number
              format: double
            profit_margin:
              type: number
              format: double
      required: [job_id]
    PriceCalculationResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        message:
          type: string
    PriceResult:
      type: object
      properties:
        item_id:
          type: string
          format: uuid
        code:
          type: string
        description_fa:
          type: string
        unit:
          type: string
        base_price:
          type: number
          format: double
        updated_price:
          type: number
          format: double
        final_price:
          type: number
          format: double
        adjustments:
          type: object
          properties:
            risk_buffer:
              type: number
              format: double
            payment_terms:
              type: number
              format: double
            profit_margin:
              type: number
              format: double
        index_snapshot:
          type: object
        calculated_at:
          type: string
          format: date-time
    PriceComparison:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PriceResult'
        summary:
          type: object
          properties:
            total_base:
              type: number
              format: double
            total_updated:
              type: number
              format: double
            total_final:
              type: number
              format: double
            avg_change_pct:
              type: number
              format: double

    # ===== Forecasting =====
    ForecastRequest:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        item_ids:
          type: array
          items:
            type: string
            format: uuid
        horizons:
          type: array
          items:
            type: integer
            enum: [1, 3, 6, 12]
          default: [1, 3, 6, 12]
        scenarios:
          type: array
          items:
            type: string
            enum: [optimistic, base, pessimistic]
          default: [optimistic, base, pessimistic]
        force_refresh:
          type: boolean
          default: false
      required: [job_id]
    ForecastResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        message:
          type: string
    ForecastResult:
      type: object
      properties:
        item_id:
          type: string
          format: uuid
        code:
          type: string
        description_fa:
          type: string
        unit:
          type: string
        current_price:
          type: number
          format: double
        forecasts:
          type: array
          items:
            $ref: '#/components/schemas/ForecastScenario'
    ForecastScenario:
      type: object
      properties:
        horizon_months:
          type: integer
        scenario:
          type: string
          enum: [optimistic, base, pessimistic]
        predicted_price:
          type: number
          format: double
        lower_bound:
          type: number
          format: double
        upper_bound:
          type: number
          format: double
        change_pct:
          type: number
          format: double
        model_type:
          type: string
          enum: [prophet, arima, lstm]
        assumptions:
          type: object
    ChapterForecastAggregate:
      type: object
      properties:
        chapter_id:
          type: string
          format: uuid
        chapter_code:
          type: string
        chapter_name:
          type: string
        item_count:
          type: integer
        horizons:
          type: array
          items:
            $ref: '#/components/schemas/ChapterHorizonAggregate'
    ChapterHorizonAggregate:
      type: object
      properties:
        horizon_months:
          type: integer
        scenarios:
          type: object
          properties:
            optimistic:
              $ref: '#/components/schemas/AggregateStats'
            base:
              $ref: '#/components/schemas/AggregateStats'
            pessimistic:
              $ref: '#/components/schemas/AggregateStats'
    AggregateStats:
      type: object
      properties:
        mean:
          type: number
          format: double
        median:
          type: number
          format: double
        total:
          type: number
          format: double
        std_dev:
          type: number
          format: double
        confidence:
          type: string

    # ===== Exports =====
...
paths:
...
  /prices/recalculate:
    post:
      tags: [Price Calculation]
      summary: Trigger price recalculation with market indices
      operationId: recalculatePrices
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceCalculationRequest'
      responses:
        '202':
          description: Calculation job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceCalculationResponse'
  /prices/{job_id}:
    get:
      tags: [Price Calculation]
      summary: Get calculated prices for all items
      operationId: getCalculatedPrices
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Price calculation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceComparison'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /prices/{job_id}/comparison:
    get:
      tags: [Price Calculation]
      summary: Get price comparison (base vs updated vs final)
      operationId: getPriceComparison
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Price comparison data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceComparison'

  # ===== Forecasting =====
  /forecasts/generate:
    post:
      tags: [Forecasting]
      summary: Generate price forecasts
      operationId: generateForecasts
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForecastRequest'
      responses:
        '202':
          description: Forecast job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastResponse'
  /forecasts/{job_id}:
    get:
      tags: [Forecasting]
      summary: Get forecast results
      operationId: getForecasts
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: horizon
          in: query
          schema:
            type: integer
            enum: [1, 3, 6, 12]
        - name: scenario
          in: query
          schema:
            type: string
            enum: [optimistic, base, pessimistic]
      responses:
        '200':
          description: Forecast results
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ForecastResult'
                  chapter_aggregates:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChapterForecastAggregate'
  /forecasts/{job_id}/chart-data:
    get:
      tags: [Forecasting]
      summary: Get chart-ready forecast data for visualization
      operationId: getForecastChartData
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: item_ids
          in: query
          schema:
            type: array
            items:
              type: string
              format: uuid
      responses:
        '200':
          description: Chart data for Recharts/Chart.js
          content:
            application/json:
              schema:
                type: object
                properties:
                  series:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        data:
                          type: array
                          items:
                            type: object
                            properties:
                              x:
                                type: string
                              y:
                                type: number
                              y_min:
                                type: number
                              y_max:
                                type: number

  # ===== Exports =====
...